third_party/blink/web_tests/svg/dynamic-updates/SVGPatternElement-dom-patternTransform-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'patternTransform' attribute of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.getAttribute('patternTransform') is "matrix(1,0,0,1,0,0)"
PASS patternElement.getAttribute('patternTransform') is "matrix(4,0,0,1,0,0)"
PASS successfullyParsed is true

TEST COMPLETE
